# Conflicts:
#	src/main/java/com/project/echoproject/controller/AdminController.java
#	src/main/resources/templates/admin/layout_admin.html
#	src/main/resources/templates/notice_detail.html
#	src/main/resources/templates/notice_list.html

+++ src/main/java/com/project/echoproject/service/MypageServiceImpl.java
@@ -6,22 +6,24 @@ import com.project.echoproject.entity.SiteUser;
 import com.project.echoproject.repository.SiteUserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.Optional;
 
 @Service  // 이 클래스가 서비스 계층의 빈으로 등록되도록 표시합니다.
 public class MypageServiceImpl implements MypageService {
 
     private final SiteUserRepository userRepository;
     private final ImageService imageService;
+    private final PasswordEncoder passwordEncoder;
 
     @Autowired  // 생성자를 통해 의존성을 주입합니다.
-    public MypageServiceImpl(SiteUserRepository userRepository, ImageService imageService) {
+    public MypageServiceImpl(SiteUserRepository userRepository, ImageService imageService, PasswordEncoder passwordEncoder) {
         this.userRepository = userRepository;
         this.imageService = imageService;
+        this.passwordEncoder = passwordEncoder;
     }
 
     // 주어진 사용자 ID로 사용자를 검색합니다.
@@ -32,32 +34,27 @@ public class MypageServiceImpl implements MypageService {
     }
 
     // 주어진 사용자 ID에 해당하는 사용자의 정보를 업데이트합니다.
+    @Override
     public void updateUser(String userId, SiteUser updatedUser, MultipartFile file) throws IOException {
         SiteUser updateInfo = userRepository.findByUserId(userId)
                 .orElseThrow(() -> new IllegalArgumentException("User not found"));
 
-        // 사용자 정보를 업데이트합니다.
         updateInfo.setUserName(updatedUser.getUserName());
+        updateInfo.setNickName(updatedUser.getNickName());
         updateInfo.setEmail(updatedUser.getEmail());
         updateInfo.setPhoneNum(updatedUser.getPhoneNum());
         updateInfo.setZipcode(updatedUser.getZipcode());
         updateInfo.setStreetaddr(updatedUser.getStreetaddr());
         updateInfo.setDetailaddr(updatedUser.getDetailaddr());
 
-        // 파일이 비어 있지 않은 경우 프로필 이미지를 업데이트합니다.
         if (file != null && !file.isEmpty()) {
-            // 기존 프로필 이미지가 있다면 삭제합니다.
-            if (updateInfo.getProfileImage() != null) {
+            Image newImage = imageService.saveImage(file);
+            if (updateInfo.getProfileImage() != null && !updateInfo.getProfileImage().isDefaultImage()) {
                 imageService.deleteImage(updateInfo.getProfileImage());
             }
-
-            // 새 이미지를 저장합니다.
-            Image newImage = imageService.saveImage(file);
             updateInfo.setProfileImage(newImage);
-            newImage.setSiteUser(updateInfo);  // 이미지 엔티티와 사용자 엔티티를 연관시킵니다.
         }
 
-        // 업데이트된 사용자 정보를 저장합니다.
         userRepository.save(updateInfo);
     }
 
@@ -76,7 +73,7 @@ public class MypageServiceImpl implements MypageService {
 
         if (updatedUser.getProfileImage() != null && !updatedUser.getProfileImage().isEmpty()) {
             Image newImage = imageService.saveImage(updatedUser.getProfileImage());
-            if (user.getProfileImage() != null) {
+            if (user.getProfileImage() != null && !user.getProfileImage().isDefaultImage()) {
                 imageService.deleteImage(user.getProfileImage());
             }
             user.setProfileImage(newImage);
@@ -84,7 +81,6 @@ public class MypageServiceImpl implements MypageService {
 
         userRepository.save(user);
     }
-
 
     // 주어진 사용자 ID에 해당하는 사용자를 삭제합니다.
     @Transactional  // 트랜잭션 범위 내에서 실행되도록 표시합니다.
